MSS Stream Modification && FFmpeg Pipping:
Moved the mss screen capturing function into a separate program/file.
Utilised the python sys module to write/output the captured frame with the stdout.buffer wrapper.
Used python's subprocess to pipe in the frame output from mss in the FFmpeg encoding function.
This enables the Graceful termination of the running FFmpeg process as FFmpeg was started using the .execute FFmpeg function.