Created classes for recorder.py and screen_capture.py:
Created two classes in the recorder.py program.
Setup an Encoder class with a method that helps detect users system hardware for setting ffmpeg encoder.
Setup a ScreenRecorder class with a constructor that takes output file name, screen dimension, selected encoder etc..
Created ffmpeg_encoder as a method in the ScreenRecorder class that encodes the video && audio from MSS and dshow respectively.
Setup a start_recording method that is responsible for starting the recording by setting up a thread for the ffmpeg_encoder method.
Setup a FrameCapture class in the screen_capture program with a screen_capture method that captures users screen and sends each captured frame to the recorder.py program using the sys.stdout.buffer module.
Created a requirements.txt file containing the necessary libraries to run the program.
Added an audio codec  for the dshow audio input.