Automation Startup Check on Launch:
Added a logic to check if the automation checkBox in Settings is True and if so, start the automation script on application launch.
Modified the settings module to always read the latest user saving path from config.json using a new method.
Removed all print statements used for debugging in every module/script.
Added an option to stop the automation script before starting a manual recording and to restart the automation script when manual recording is stopped.